refactor: restructure lib folders for clarity and MVVM
