+++ lib/presentation/pages/edit_goal/edit_goal_page.dart
@@ -3,8 +3,8 @@ import 'package:provider/provider.dart';
 
 import 'package:haenaedda/gen_l10n/app_localizations.dart';
 import 'package:haenaedda/theme/buttons.dart';
-import 'package:haenaedda/presentation/goal_calendar/goal_edit_result.dart';
-import 'package:haenaedda/presentation/settings/handlers/discard_changes_handler.dart';
+import 'package:haenaedda/presentation/pages/edit_goal/goal_edit_result.dart';
+import 'package:haenaedda/presentation/handlers/discard_changes_handler.dart';
 import 'package:haenaedda/presentation/widgets/discard_aware_close_button.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
