+++ lib/presentation/pages/goal_calendar/goal_calendar_header.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:haenaedda/model/goal.dart';
-import 'package:haenaedda/presentation/goal_calendar/month_navigation_bar.dart';
+import 'package:haenaedda/presentation/pages/goal_calendar/month_navigation_bar.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
 
 class GoalCalendarHeader extends StatelessWidget {
