+++ lib/presentation/pages/launcher/launcher_page.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:haenaedda/presentation/goal_calendar/goal_calendar_page.dart';
-import 'package:haenaedda/presentation/launcher/add_first_goal_flow.dart';
+import 'package:haenaedda/presentation/pages/goal_calendar/goal_calendar_page.dart';
+import 'package:haenaedda/presentation/pages/launcher/add_first_goal_flow.dart';
 import 'package:haenaedda/presentation/widgets/loading_indicator.dart';
 import 'package:haenaedda/presentation/view_models/goal_view_models.dart';
