+++ lib/presentation/pages/settings/settings_bottom_modal.dart
@@ -5,8 +5,8 @@ import 'package:haenaedda/gen_l10n/app_localizations.dart';
 import 'package:haenaedda/model/goal.dart';
 import 'package:haenaedda/model/goal_setting_action.dart';
 import 'package:haenaedda/model/reset_type.dart';
-import 'package:haenaedda/presentation/settings/handlers/reset_goal_handler.dart';
-import 'package:haenaedda/presentation/settings/neumorphic_settings_tile.dart';
+import 'package:haenaedda/presentation/handlers/reset_goal_handler.dart';
+import 'package:haenaedda/presentation/pages/settings/neumorphic_settings_tile.dart';
 import 'package:haenaedda/presentation/widgets/bottom_right_button.dart';
 
 class SettingsBottomModal extends StatelessWidget {
