+++ lib/presentation/view_models/goal_view_models.dart
@@ -6,7 +6,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 import 'package:haenaedda/constants/storage_keys.dart';
 import 'package:haenaedda/model/goal.dart';
-import 'package:haenaedda/utils/extensions/iterable_extensions.dart';
+import 'package:haenaedda/extensions/iterable_extensions.dart';
 import 'package:haenaedda/presentation/view_models/goal_result.dart';
 
 class GoalViewModel extends ChangeNotifier {
