+++ lib/theme/app_theme.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:haenaedda/theme/dark_theme.dart';
-import 'package:haenaedda/theme/light_theme.dart';
+import 'package:haenaedda/theme/colors/dark_theme.dart';
+import 'package:haenaedda/theme/colors/light_theme.dart';
 
 class AppTheme {
   static ThemeData get light => lightTheme;
