+++ lib/theme/colors/color_scheme_dark.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:haenaedda/theme/app_colors.dart';
+import 'package:haenaedda/theme/colors/app_colors.dart';
 
 const darkColorScheme = ColorScheme(
   brightness: Brightness.dark,
