+++ lib/theme/colors/light_theme.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:haenaedda/theme/app_colors.dart';
+import 'package:haenaedda/theme/colors/app_colors.dart';
 import 'package:haenaedda/theme/app_typography.dart';
-import 'package:haenaedda/theme/color_scheme_light.dart';
-import 'package:haenaedda/theme/surface_container_colors_light.dart';
+import 'package:haenaedda/theme/colors/color_scheme_light.dart';
+import 'package:haenaedda/theme/colors/surface_container_colors_light.dart';
 
 final ThemeData lightTheme = ThemeData(
   brightness: Brightness.light,
